payments_manager_django/requirements.txt
Django==3.1.3
django-environ==0.4.5
python-dotenv==0.19.0

# DRF
djangorestframework==3.12.2
django-cors-headers==3.5.0
coreapi==2.3.3

# DB
mysqlclient==2.0.1
django-mysql==3.9

# AMQP
pika==1.1.0

# Development
ipdb==0.13.7

# Filters
django-filter==2.4.0

# Tests
coverage==5.5

# PEP8
black==20.8b1

# CPF
brazilnum==0.8.8

# Swagger
drf-yasg==1.20.0
flex==6.14.1
swagger-spec-validator==2.7.3